Rename HW 4 Work.ipynb to VLE liquid composition
The goal of this file was to determine the liquid compositions of 1-propanol and 1-butanol in Vapor-Liquid Equilibrium, using Antoine's Equation and Raoult's Law. Initial conditions were provided, such as the vapor compositions and the temperature. Since pressure was also unknown, scipy.optimize.fsolve was used to more efficiently solve the system of equations.